requirements.txt
pillow
typer
python-barcode
colorama